. rpcf-dev.install.in, configure.ac, monitor/Makefile.am, appmoncli/Makefile.am cpp/Makefile.am, maincli.cpp, maincli2.cpp, AppManagercli.cpp, AppManagercli.h, AppMonitorcli.cpp, AppMonitorcli.h, SimpleAuthcli.cpp, SimpleAuthcli.h: move client code from appmon to appmoncli/cpp, and build as shared library. updcfg.py.in: made sure the SimpleAuth and AppManager clients connect to server via IPC currently. defines.h: added HANDLE the namespace to avoid conflict with other HANDLE definition, such as windows HANDLE definition.
