+++ configure.ac
@@ -432,8 +432,10 @@ AC_CONFIG_FILES([Makefile
                  monitor/oinit/Makefile
                  monitor/oinit/setup.py
                  monitor/oinit/oinit.spec
-                 monitor/appmon/Makefile
                  monitor/usermgmt/Makefile
+                 monitor/appmon/Makefile
+                 monitor/appmoncli/Makefile
+                 monitor/appmoncli/cpp/Makefile
                  monitor/Makefile
                  docs/Makefile
                  rpcf.pc ])
