+++ monitor/appmon/Makefile.am
@@ -1,14 +1,14 @@
 link_dir=$(top_srcdir)/combase/
 include_ipcdir=$(top_srcdir)/ipc
 
-appmon_PROGRAMS=appmonsvr appmoncli
+appmon_PROGRAMS=appmonsvr
 appmonsvr_SOURCES=mainsvr.cpp AppMonitorsvr.cpp AppMonitorsvr.h monconst.h \
 SimpleAuthsvr.cpp SimpleAuthsvr.h AppManagersvr.cpp AppManagersvr.h
 if ENABLE_FUSE3
 appmonsvr_SOURCES+=monfuse.cpp
 endif
-appmoncli_SOURCES=maincli.cpp AppMonitorcli.cpp AppMonitorcli.h  monconst.h\
-SimpleAuthcli.cpp SimpleAuthcli.h AppManagercli.cpp AppManagercli.h
+#appmoncli_SOURCES=maincli.cpp AppMonitorcli.cpp  monconst.h\
+#SimpleAuthcli.cpp AppManagercli.cpp 
 
 appmonsvr_LDADD= ./libappmon.a \
 ../registry/libregfs.la ../../ipc/libipc.la $(COMMON_LINK_FUSE3) \
@@ -16,7 +16,7 @@ appmonsvr_LDADD= ./libappmon.a \
  -L$(abs_top_srcdir)/combase -L$(top_srcdir)/ipc \
  -L$(top_srcdir)/3rd_party 
 
-appmoncli_LDADD=$(appmonsvr_LDADD)
+#appmoncli_LDADD=$(appmonsvr_LDADD)
 
 appmonconf_DATA=appmondesc.json appmon.ridl
 appmonconfdir=$(sysconfdir)/rpcf
@@ -37,23 +37,24 @@ dist_appmonscr_SCRIPTS=initappreg.sh appfuncs.sh
 appmonscrdir=$(bindir)/rpcf
 endif
 
-appmoncli_CPPFLAGS=$(appmonsvr_CPPFLAGS)
-appmoncli_LDFLAGS=$(appmonsvr_LDFLAGS)
+#appmoncli_CPPFLAGS=$(appmonsvr_CPPFLAGS)
+#appmoncli_LDFLAGS=$(appmonsvr_LDFLAGS)
 
 EXTRA_DIST=appmon.ridl
 
 noinst_LIBRARIES=libappmon.a
-libappmon_a_SOURCES = appmon.cpp appmonstruct.cpp IAppStorecli.cpp IAppStoresvr.cpp IAppMonitorcli.cpp IAppMonitorsvr.cpp IDataProducercli.cpp IDataProducersvr.cpp IUserManagercli.cpp IUserManagersvr.cpp appmondesc.json IUserManagersvr.h IUserManagercli.h IDataProducersvr.h IDataProducercli.h IAppMonitorsvr.h IAppMonitorcli.h IAppStoresvr.h IAppStorecli.h appmon.h appmonstruct.h 
+libappmon_a_SOURCES = appmon.cpp appmonstruct.cpp IAppStoresvr.cpp IAppMonitorsvr.cpp IDataProducersvr.cpp IUserManagersvr.cpp appmondesc.json IUserManagersvr.h IDataProducersvr.h IAppMonitorsvr.h IAppStoresvr.h appmon.h appmonstruct.h 
 libappmon_a_CPPFLAGS= $(appmonsvr_CPPFLAGS)
 libappmon_a_HEADERS=
+
+libappmon_adir=$(includedir)/rpcf/appmon
 #libappmon_la_LIBADD= ../registry/libregfs.la ../../ipc/libipc.la ../../combase/libcombase.la -lpthread -ldl $(COMMON_LINK) -L../../3rd_party
 #libappmon_la_LDFLAGS= -version-info @VERSION2@
 #libappmon_ladir=$(libdir)
-libappmon_adir=$(includedir)/rpcf/appmon
 
-headers=$(libappmon_a_HEADERS) AppMonitorsvr.h AppMonitorcli.h SimpleAuthcli.h SimpleAuthsvr.h AppManagercli.h AppManagersvr.h monconst.h
+headers=$(libappmon_a_HEADERS) AppMonitorsvr.h SimpleAuthsvr.h AppManagersvr.h monconst.h
 
-sources=$(libappmon_a_SOURCES) AppMonitorsvr.cpp mainsvr.cpp maincli.cpp AppMonitorcli.cpp SimpleAuthcli.cpp SimpleAuthsvr.cpp AppManagercli.cpp AppManagersvr.cpp 
+sources=$(libappmon_a_SOURCES) AppMonitorsvr.cpp mainsvr.cpp SimpleAuthsvr.cpp AppManagersvr.cpp 
 
 $(sources):$(headers)
 
@@ -69,11 +70,10 @@ appmon.ridl:
     fi || true; \
 	if [ -f Makefile ]; then cp Makefile Makefile_; fi; \
     if [ -f ../../ridl/.libs/ridlc ]; then \
-        LD_LIBRARY_PATH=../../combase/.libs:$${LD_LIBRARY_PATH} ../../ridl/.libs/ridlc -sO . ./appmon.ridl; \
+        LD_LIBRARY_PATH=../../combase/.libs:$${LD_LIBRARY_PATH} ../../ridl/.libs/ridlc --server -sO . ./appmon.ridl; \
     else \
         echo ridlc not found. Please build the whole project from the root directory first; exit 1; fi; \
 	if [ -f Makefile_ ]; then mv Makefile_ Makefile; fi; \
-	if [ -f ./mktpl ];then rm -f mktpl mktplsvr mktplcli odesctpl2.json drvtpl.json;fi || true; \
-    ls
+	if [ -f ./mktpl ];then rm -f mktpl mktplsvr mktplcli odesctpl2.json drvtpl.json;fi || true;
 
 .PHONY: appmon.ridl
