--- monitor/appmon/maincli.cpp
@@ -1,158 +1,0 @@
-// GENERATED BY RIDLC, MAKE SURE TO BACKUP BEFORE RUNNING RIDLC AGAIN
-// Copyright (C) 2024  zhiming <woodhead99@gmail.com>
-// This program can be distributed under the terms of the GNU GPLv3.
-// ../../ridl/.libs/ridlc -sO . ./appmon.ridl 
-#include "rpc.h"
-#include "proxy.h"
-using namespace rpcf;
-#include "stmport.h"
-#include "fastrpc.h"
-#include "AppMonitorcli.h"
-#include "SimpleAuthcli.h"
-
-ObjPtr g_pIoMgr;
-
-FactoryPtr InitClassFactory()
-{
-    BEGIN_FACTORY_MAPS;
-
-    INIT_MAP_ENTRYCFG( CAppMonitor_CliImpl );
-    INIT_MAP_ENTRYCFG( CAppMonitor_CliSkel );
-    INIT_MAP_ENTRYCFG( CAppMonitor_ChannelCli );
-    // INIT_MAP_ENTRYCFG( CSimpleAuth_CliImpl );
-    // INIT_MAP_ENTRYCFG( CSimpleAuth_CliSkel );
-    // INIT_MAP_ENTRYCFG( CSimpleAuth_ChannelCli );
-    
-    INIT_MAP_ENTRY( KeyValue );
-    
-    END_FACTORY_MAPS;
-}
-
-extern "C"
-gint32 DllLoadFactory( FactoryPtr& pFactory )
-{
-    pFactory = InitClassFactory();
-    if( pFactory.IsEmpty() )
-        return -EFAULT;
-    return STATUS_SUCCESS;
-}
-
-gint32 InitContext()
-{
-    gint32 ret = CoInitialize( 0 );
-    if( ERROR( ret ) )
-        return ret;
-    do{
-        // load class factory for 'appmon'
-        FactoryPtr p = InitClassFactory();
-        ret = CoAddClassFactory( p );
-        if( ERROR( ret ) )
-            break;
-        
-        CParamList oParams;
-        oParams.Push( "appmoncli" );
-
-        // adjust the thread number if necessary
-        oParams[ propMaxIrpThrd ] = 0;
-        oParams[ propMaxTaskThrd ] = 1;
-        
-        ret = g_pIoMgr.NewObj(
-            clsid( CIoManager ), 
-            oParams.GetCfg() );
-        if( ERROR( ret ) )
-            break;
-
-        CIoManager* pSvc = g_pIoMgr;
-        ret = pSvc->Start();
-        if( ERROR( ret ) )
-            break;
-    }while( 0 );
-
-    return ret;
-}
-
-gint32 DestroyContext()
-{
-    IService* pSvc = g_pIoMgr;
-    if( pSvc != nullptr )
-    {
-        pSvc->Stop();
-        g_pIoMgr.Clear();
-    }
-
-    CoUninitialize();
-    DebugPrintEx( logErr, 0,
-        "#Leaked objects is %d",
-        CObjBase::GetActCount() );
-    return STATUS_SUCCESS;
-}
-
-gint32 maincli(
-    CAppMonitor_CliImpl* pIf,
-    int argc, char** argv );
-
-int main( int argc, char** argv )
-{
-    gint32 ret = 0;
-    do{
-        std::string strDesc = "./appmondesc.json";
-        ret = InitContext();
-        if( ERROR( ret ) )
-            break;
-        
-        CRpcServices* pSvc = nullptr;
-        InterfPtr pIf;
-        std::vector< InterfPtr > vecIfs;
-        do{
-            CParamList oParams;
-            oParams[ propIoMgr ] = g_pIoMgr;
-            
-            ret = CRpcServices::LoadObjDesc(
-                strDesc, "AppMonitor",
-                false, oParams.GetCfg() );
-            if( ERROR( ret ) )
-                break;
-            ret = pIf.NewObj(
-                clsid( CAppMonitor_CliImpl ),
-                oParams.GetCfg() );
-            if( ERROR( ret ) )
-                break;
-            pSvc = pIf;
-            ret = pSvc->Start();
-            vecIfs.push_back( pIf );
-            if( ERROR( ret ) )
-                break;
-            while( pSvc->GetState()== stateRecovery )
-                sleep( 1 );
-            
-            if( pSvc->GetState() != stateConnected )
-            {
-                ret = ERROR_STATE;
-                break;
-            }
-
-        }while( 0 );
-        
-        if( SUCCEEDED( ret ) )
-            ret = maincli(
-                vecIfs[ 0 ],
-                argc, argv );
-            
-        // Stopping the objects
-        for( auto& pInterf : vecIfs )
-            pInterf->Stop();
-    }while( 0 );
-
-    DestroyContext();
-    return ret;
-}
-
-//-----Your code begins here---
-gint32 maincli(
-    CAppMonitor_CliImpl* pIf,
-    int argc, char** argv )
-{
-    //-----Adding your code here---
-    return STATUS_SUCCESS;
-}
-
