+++ monitor/appmoncli/cpp/SimpleAuthcli.h
@@ -1,7 +1,7 @@
 // GENERATED BY RIDLC, MAKE SURE TO BACKUP BEFORE RUNNING RIDLC AGAIN
 // Copyright (C) 2024  zhiming <woodhead99@gmail.com>
 // This program can be distributed under the terms of the GNU GPLv3.
-// ../../ridl/.libs/ridlc -sO . ./appmon.ridl 
+// ../../../ridl/.libs/ridlc --client -slO . ../../appmon/appmon.ridl 
 #pragma once
 #include "appmon.h"
